[feature] #kod-103 - add breadcrumb and biund it to navigation
refactor auth service to save entityId in local storage
add breadcrumb actions and reducer
add higher level navigation actions to be used in history service
add breadcrumb param functions in param service
clean some typo

+++ src/scripts/store/persistenceStore.js
@@ -28,20 +28,25 @@ export default function persistenceHandler(next) {
 
         const storeState = store.getState()
 
-        storageService.put('locale', storeState.prefs.locale)
-        storageService.put('theme', storeState.prefs.theme)
-        storageService.put('navigation', storeState.prefs.navigation)
-        storageService.put('navigation', storeState.prefs.navigation)
-        const projectConfigId = storeState.projectConfig.id
-        if (projectConfigId) {
-          storageService.put('projectConfigId', projectConfigId)
-          const project = storeState.projectConfig.project
-          if (project) {
-            storageService.put('projectId', project.id)
+        if (storeState) {
+          if (storeState.prefs) {
+            storageService.put('locale', storeState.prefs.locale)
+            storageService.put('theme', storeState.prefs.theme)
+            storageService.put('navigation', storeState.prefs.navigation)
           }
-        } else {
-          storageService.remove('projectConfigId')
-          storageService.remove('projectId')
+          if (storeState.projectConfig) {
+            const projectConfigId = storeState.projectConfig.id
+            if (projectConfigId) {
+              storageService.put('projectConfigId', projectConfigId)
+              const project = storeState.projectConfig.project
+              if (project) {
+                storageService.put('projectId', project.id)
+              }
+            } else {
+              storageService.remove('projectConfigId')
+              storageService.remove('projectId')
+            }
+          }
         }
 
         return action
